+++ traefik/configs/dynamic/internal.yaml
@@ -9,9 +9,9 @@ tls:
 
 http:
   middlewares:
-    local-middleware: # sources are located in traefik/plugins/middleware directory
+    docker-info-middleware: # sources are located in traefik/plugins/docker_info directory
       plugin:
-        local-middleware: {} # plugin and middleware names are same
+        docker-info: {} # plugin and docker_info names are same
 
     traefik-dashboard-cors:
       headers:
@@ -28,4 +28,4 @@ http:
       tls: {} # force TLS usage
       middlewares:
         - traefik-dashboard-cors # allow limited CORS for traefik dashboard API
-        - local-middleware # use our local middleware together with the traefik dashboard API
+        - docker-info-middleware # use our local docker_info together with the traefik dashboard API
