+++ traefik/configs/traefik.yaml
@@ -7,8 +7,8 @@ global:
 
 experimental:
   localPlugins:
-    local-middleware: # sources are located in traefik/plugins/middleware directory
-      moduleName: middleware
+    docker-info: # sources are located in traefik/plugins/docker_info directory
+      moduleName: docker_info
 
 serversTransport: # https://doc.traefik.io/traefik/routing/services/#serverstransport_1
   insecureSkipVerify: true
