+++ traefik/plugins/docker_info/.traefik.yml
@@ -1,0 +1,7 @@
+displayName: docker-info
+summary: Docker info plugin
+type: middleware
+
+import: docker_info
+
+testData: {}
