+++ traefik/plugins/docker_info/go.mod
@@ -1,0 +1,3 @@
+module docker_info
+
+go 1.18
