--- traefik/plugins/middleware/.traefik.yml
@@ -1,7 +1,0 @@
-displayName: local-middleware
-summary: Local middleware plugin
-type: middleware
-
-import: middleware
-
-testData: {}
