--- traefik/plugins/middleware/go.mod
@@ -1,3 +1,0 @@
-module middleware
-
-go 1.18
