--- traefik/plugins/middleware/plugin_test.go
@@ -1,39 +1,0 @@
-package middleware_test
-
-import (
-	"context"
-	"net/http"
-	"net/http/httptest"
-	"testing"
-
-	"middleware"
-)
-
-func TestPlugin_ServeHTTP(t *testing.T) {
-	t.Parallel()
-
-	var (
-		req, _ = http.NewRequest(http.MethodGet, "http://test/local-middleware/containers/list", http.NoBody)
-		rr     = httptest.NewRecorder()
-		next   = http.HandlerFunc(func(rw http.ResponseWriter, req *http.Request) { rw.WriteHeader(http.StatusOK) })
-	)
-
-	plugin, err := middleware.New(
-		context.Background(),
-		next,
-		middleware.CreateConfig(),
-		"",
-	)
-
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	plugin.ServeHTTP(rr, req)
-
-	if rr.Code != http.StatusOK {
-		t.Errorf("expected status code %d, got %d", http.StatusOK, rr.Code)
-	}
-
-	t.Log(rr.Body.String()) // TODO: for debugging, remove later
-}
